Change the way description is filled: instead of using page.type, we paste the description
This solves a new issue that arose: characters are getting out of order with page.type, so the way of filling description is changed from page.type to pasting the description.

Note that it is important to first focus the description textbox, then invoke selectAll for the textbox before pasting the description.